metadata api, nav favicon, upgrade
