+++ package.json
@@ -9,39 +9,39 @@
     "lint": "next lint"
   },
   "dependencies": {
-    "@axe-core/react": "^4.7.2",
-    "@navikt/aksel-icons": "^4.1.2",
-    "@navikt/ds-css": "^5.0.3",
+    "@axe-core/react": "^4.7.3",
+    "@navikt/aksel-icons": "^5.2.1",
+    "@navikt/ds-css": "^5.2.1",
     "@navikt/ds-icons": "^3.4.3",
-    "@navikt/ds-react": "^5.0.3",
-    "@navikt/ds-tokens": "^4.1.2",
+    "@navikt/ds-react": "^5.2.1",
+    "@navikt/ds-tokens": "^5.2.1",
     "amplitude-js": "^8.21.9",
     "classnames": "^2.3.2",
-    "framer-motion": "^8.2.0",
-    "next": "13.4.4",
+    "framer-motion": "^10.16.2",
+    "next": "13.4.19",
     "query-string": "^8.1.0",
     "react": "18.2.0",
     "react-dom": "18.2.0",
-    "react-hook-form": "^7.44.3",
+    "react-hook-form": "^7.45.4",
     "react-hotjar": "^6.1.0",
-    "react-intersection-observer": "^9.4.4",
+    "react-intersection-observer": "^9.5.2",
     "styled-components": "^6.0.7",
-    "swr": "^2.1.5",
-    "zustand": "^4.3.8"
+    "swr": "^2.2.2",
+    "zustand": "^4.4.1"
   },
   "devDependencies": {
-    "@trivago/prettier-plugin-sort-imports": "^4.1.1",
+    "@trivago/prettier-plugin-sort-imports": "^4.2.0",
     "@types/amplitude-js": "^8.16.2",
     "@types/deep-equal": "^1.0.1",
-    "@types/node": "18.11.9",
-    "@types/react": "18.2.9",
-    "@types/react-dom": "18.2.4",
-    "eslint": "8.42.0",
-    "eslint-config-next": "13.4.4",
+    "@types/node": "20.5.7",
+    "@types/react": "18.2.21",
+    "@types/react-dom": "18.2.7",
+    "eslint": "8.48.0",
+    "eslint-config-next": "13.4.19",
     "eslint-plugin-jsx-a11y": "^6.7.1",
-    "prettier": "^2.8.8",
-    "sass": "^1.63.2",
-    "sharp": "^0.32.4",
-    "typescript": "4.8.4"
+    "prettier": "^3.0.3",
+    "sass": "^1.66.1",
+    "sharp": "^0.32.5",
+    "typescript": "5.2.2"
   }
 }
